backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
pydantic[email]
passlib
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
pdfplumber
